+++ hsmd/Makefile
@@ -31,6 +31,8 @@ HSMD_COMMON_OBJS :=				\
 	common/msg_queue.o			\
 	common/node_id.o			\
 	common/permute_tx.o			\
+	common/psbt_open.o			\
+	common/pseudorand.o			\
 	common/setup.o				\
 	common/status.o				\
 	common/status_wire.o			\
@@ -39,11 +41,6 @@ HSMD_COMMON_OBJS :=				\
 	common/utils.o				\
 	common/utxo.o				\
 	common/version.o
-
-ifeq ($(EXPERIMENTAL_FEATURES),1)
-HSMD_COMMON_OBJS += common/psbt_open.o		\
-	common/pseudorand.o
-endif
 
 lightningd/lightning_hsmd: $(HSMD_OBJS) $(HSMD_COMMON_OBJS) $(BITCOIN_OBJS) $(WIRE_OBJS)
