+++ onchaind/Makefile
@@ -53,6 +53,8 @@ ONCHAIND_COMMON_OBJS :=				\
 	common/onionreply.o			\
 	common/peer_billboard.o			\
 	common/permute_tx.o			\
+	common/psbt_open.o			\
+	common/pseudorand.o			\
 	common/setup.o				\
 	common/status.o				\
 	common/status_wire.o			\
@@ -62,11 +64,6 @@ ONCHAIND_COMMON_OBJS :=				\
 	common/utxo.o				\
 	common/version.o			\
 	common/wallet.o
-
-ifeq ($(EXPERIMENTAL_FEATURES),1)
-ONCHAIND_COMMON_OBJS += common/psbt_open.o	\
-	common/pseudorand.o
-endif
 
 lightningd/lightning_onchaind: $(ONCHAIND_OBJS) $(WIRE_ONION_OBJS) $(ONCHAIND_COMMON_OBJS) $(WIRE_OBJS) $(BITCOIN_OBJS) $(HSMD_CLIENT_OBJS)
